Refit function for cv_quantile_lasso object

Refit lasso solutions at a new set of quantile levels, given an existing cv_quantile_lasso object.

refit_quantile_lasso(
  obj,
  x,
  y,
  tau_new = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99),
  weights = NULL,
  no_pen_vars = NULL,
  intercept = TRUE,
  standardize = TRUE,
  noncross = FALSE,
  x0 = NULL,
  lp_solver = NULL,
  time_limit = NULL,
  warm_starts = NULL,
  params = NULL,
  transform = NULL,
  inv_trans = NULL,
  jitter = NULL,
  verbose = FALSE
)

Arguments

obj
The cv_quantile_lasso object to start from.

x
Matrix of predictors.

y
Vector of responses.

tau_new
Vector of new quantile levels at which to fit new solutions. Default is a sequence of 23 quantile levels from 0.01 to 0.99.

noncross
Should noncrossing constraints be applied? These force the estimated quantiles to be properly ordered across all quantile levels being considered. The default is FALSE. If TRUE, then noncrossing constraints are applied to the estimated quantiles at all points specified by the next argument x0.

x0
Matrix of points used to define the noncrossing constraints. Default is NULL, which means that we consider noncrossing constraints at the training points x.

verbose
Should progress be printed out to the console? Default is FALSE.

Value

A quantile_lasso object, with solutions at quantile levels tau_new.

Details

This function simply infers, for each quantile level in tau_new, a (very) roughly-CV-optimal tuning parameter value, then calls quantile_lasso at the new quantile levels and corresponding tuning parameter values. If not specified, the arguments weights, no_pen_vars, intercept, standardize, lp_solver, time_limit, warm_start, params, transform, inv_transorm, jitter are all inherited from the given cv_quantile_lasso object.